requirements.txt
python-dotenv
fastapi
uvicorn
pydantic
requests
jinja2
together